refact: move groth16 script to related mod
